fixing imports and exports of modals to cut down on problems
