+++ web/apps/labelstudio/src/components/Modal/Modal.tsx
@@ -9,33 +9,18 @@ import {
   modal as coreModal,
   confirm as coreConfirm,
   info as coreInfo,
-  type ModalProps as CoreModalProps,
-  type ConfirmProps as CoreConfirmProps,
-  type InfoProps as CoreInfoProps,
-  type ModalUpdateProps as CoreModalUpdateProps,
-} from "@humansignal/ui";
+  type ModalProps,
+  type ModalUpdateProps,
+  type ExtraProps,
+} from "@humansignal/ui/lib/modal";
 import { ApiProvider } from "../../providers/ApiProvider";
 import { AuthProvider } from "@humansignal/core/providers/AuthProvider";
 import { ConfigProvider } from "../../providers/ConfigProvider";
-import { ToastProvider } from "@humansignal/ui";
+import { ToastProvider } from "@humansignal/ui/lib/toast/toast";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { queryClient } from "../../utils/query-client";
 
-export type { ButtonProps as ButtonVariant } from "@humansignal/ui";
-
-export type ConfirmProps<T> = CoreConfirmProps<T> & ExtraProps;
-export type InfoProps<T> = CoreInfoProps<T> & ExtraProps;
-export type ModalProps<T = unknown> = CoreModalProps<T>;
-// biome-ignore lint/suspicious/noExplicitAny: Generic type parameter for backward compatibility
-export type ModalUpdate<Props extends ModalProps<any>> = CoreModalUpdateProps<any>;
-export type ModalUpdateProps<T> = CoreModalUpdateProps<T>;
-
-export type ExtraProps = {
-  unique?: string;
-  simple?: boolean;
-  onHidden?: () => void;
-  providers?: ReactElement[];
-};
+export type { ButtonProps as ButtonVariant } from "@humansignal/ui/lib/button/button";
 
 /**
  * Get the default LS providers for modals
@@ -50,49 +35,25 @@ const getDefaultProviders = (): ReactElement[] => {
   ];
 };
 
-/**
- * Create a modal with LS providers automatically injected.
- * Maintains backward compatibility with existing LS code.
- */
-export const modal = <T,>(props: ModalProps<T> & ExtraProps): ModalUpdateProps<T> => {
-  const providers = props.simple ? [] : (props.providers ?? getDefaultProviders());
+const modalTypes = {
+  modal: coreModal,
+  confirm: coreConfirm,
+  info: coreInfo,
+} as const;
 
-  return coreModal({
-    ...props,
-    providers,
-  }) as ModalUpdateProps<T>;
-};
-
-/**
- * Create a confirmation modal with LS providers.
- */
-export const confirm = <T,>(props: ConfirmProps<T>): ModalUpdateProps<T> => {
-  const providers = props.simple ? [] : (props.providers ?? getDefaultProviders());
-
-  return coreConfirm({
-    ...props,
-    providers,
-  }) as ModalUpdateProps<T>;
-};
-
-/**
- * Create an informational modal with LS providers.
- */
-export const info = <T,>(props: InfoProps<T>): ModalUpdateProps<T> => {
-  const providers = props.simple ? [] : (props.providers ?? getDefaultProviders());
-
-  return coreInfo({
-    ...props,
-    providers,
-  }) as ModalUpdateProps<T>;
+const createModal = (type: keyof typeof modalTypes) => {
+  return <T,>(props: ModalProps<T> & ExtraProps): ModalUpdateProps<T> => {
+    return modalTypes[type]({
+      simple: false,
+      providers: getDefaultProviders(),
+      ...props,
+    });
+  };
 };
 
 // Re-export Modal component and hooks
-export { Modal, useModalControls } from "@humansignal/ui";
-
-// Re-export for backward compatibility
-// Note: Modal is a React component class, not a namespace, so we can't assign to it
-// The original pattern was likely different - these functions are exported separately
-
-// Export standalone modal as default
+export const modal = createModal("modal");
+export const confirm = createModal("confirm");
+export const info = createModal("info");
 export { modal as standaloneModal };
+export { Modal, useModalControls } from "@humansignal/ui/lib/modal";
