+++ web/libs/datamanager/src/components/Common/Modal/ModalPopup.tsx
@@ -3,4 +3,4 @@
  *
  * Re-exports Modal components from @humansignal/ui for backward compatibility
  */
-export { Modal, useModalControls, type ModalProps } from "@humansignal/ui";
+export { Modal, useModalControls, type ModalProps } from "@humansignal/ui/lib/modal";
